requirements.txt
langchain
streamlit
logging
pathlib
langchain_community
python-dotenv
openai
docarray
tiktoken
pypdf
langchain_experimental
wikipedia
